Remove test and prod deployments from the main Jenkins pipeline.
The test and prod deployments require input for the build to complete.  When you don't proceed with the deployments to test and prod the build shows as canceled, which is misleading when all you wanted to do was to deploy a successful build to dev and not to test or prod.

Deployment to test and prod are a separate concern.