Update pipeline build process
First pass at eliminating unnecessary builds and deployments.

Unfortunately the code contained in a Jenkins file is unable to affect the Pipeline's top level SCM settings to take advantage of the include and exclude filters that would stop the builds from being triggered.

We've taken steps at the next available level to short-circuit the build and avoid building anything if there were no changes in the context directory for the pipeline.  A 'build' still gets kicked off, but we avoid re-building and deploying things unnecessarily.

Plus;
- Updates to code documentation.
- Update initializeProjects scrip to grant Jenkins service account access to scale deployments.